Add functions+tests to load new h5 dataframes.
