+++ antea/testdata/ring_test.h5
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:04f54f59fbff4bbb564291001a66d45c70cdd380a3f9263196e299ee408eea79
+size 27936640
